Tests run under python 3 as well now
