+++ python/tests/test_model_selection.py
@@ -12,8 +12,6 @@ from tests.pyspark_test import PySparkTest
 
 # Overwrite the sklearn GridSearch in this suite so that we can run the same
 # tests with the same parameters.
-
-
 class AllTests(PySparkTest):
     # After testing, make sure to revert sklearn to normal
     # (see _add_to_module())
@@ -86,9 +84,14 @@ def _add_to_module():
         if method_name.startswith("test_")
     ]
 
+    # These tests have been edited and moved into this repo e.g. in
+    # resource_warning_tests.py
+    refactored_tests = ["test_return_train_score_warn"]
+
     for test in all_tests:
-        test_to_add = _create_method(test)
-        setattr(AllTests, test.__name__, test_to_add)
+        if test.__name__ not in refactored_tests:
+            test_to_add = _create_method(test)
+            setattr(AllTests, test.__name__, test_to_add)
 
 
 _add_to_module()
